更新:优化安装

+++ app/Controller/InstallController.php
@@ -137,6 +137,12 @@ class InstallController extends AbstractController
             'REDIS_HOST' => request()->input("REDIS_HOST"),
         ]);
         file_put_contents(BASE_PATH."/app/CodeFec/storage/install-step.txt",date("Y-m-d H:i:s"));
+
+        return response()->redirect("/install?step=5");
+    }
+
+    public function post_step5(): \Psr\Http\Message\ResponseInterface
+    {
         $command = 'migrate';
 
         $params = ["command" => $command, "--force"];
@@ -149,12 +155,6 @@ class InstallController extends AbstractController
         $application->setAutoExit(false);
 
         $exitCode = $application->run($input, $output);
-        file_put_contents(BASE_PATH."/app/CodeFec/storage/install-step.txt",date("Y-m-d H:i:s"));
-        return response()->redirect("/install?step=5");
-    }
-
-    public function post_step5(): \Psr\Http\Message\ResponseInterface
-    {
         AdminUser::query()->create([
             'email' => request()->input("email"),
             'username' => request()->input("username"),
@@ -168,33 +168,9 @@ class InstallController extends AbstractController
         if(!is_dir(BASE_PATH."/app/CodeFec/storage")){
             exec("mkdir ".BASE_PATH."/app/CodeFec/storage");
         }
-        file_put_contents(BASE_PATH."/app/CodeFec/storage/install.lock",date("Y-m-d H:i:s"));
-        $params = ["command" => "CodeFec:PluginsComposerInstall"];
-
-        $input = new ArrayInput($params);
-        $output = new NullOutput();
-
-        $container = \Hyperf\Utils\ApplicationContext::getContainer();
-
-        /** @var Application $application */
-        $application = $container->get(\Hyperf\Contract\ApplicationInterface::class);
-        $application->setAutoExit(false);
-
-        // 这种方式: 不会暴露出命令执行中的异常, 不会阻止程序返回
-        $exitCode = $application->run($input, $output);
-        $params = ["command" => "CodeFec:PluginsComposerInstall"];
-
-        $input = new ArrayInput($params);
-        $output = new NullOutput();
-
-        $container = \Hyperf\Utils\ApplicationContext::getContainer();
-
-        /** @var Application $application */
-        $application = $container->get(\Hyperf\Contract\ApplicationInterface::class);
-        $application->setAutoExit(false);
-
-        // 这种方式: 不会暴露出命令执行中的异常, 不会阻止程序返回
-        $exitCode = $application->run($input, $output);
+        if(!file_exists(BASE_PATH."/app/CodeFec/storage/install.lock")){
+            file_put_contents(BASE_PATH."/app/CodeFec/storage/install.lock",date("Y-m-d H:i:s"));
+        }
         return redirect()->url("/admin")->go();
     }
 }
